#292: Tune report generator to use hardware efficiently and increase speed
#293: Error limiter is not fully effective

+++ src/main/java/com/xceptance/xlt/engine/XltJavaScriptEngine.java
@@ -103,7 +103,7 @@ public final class XltJavaScriptEngine extends JavaScriptEngine
             {
                 final long runTimeInNS = System.nanoTime() - start;
 
-                customData.setRunTime(runTimeInNS / 1000000L);
+                customData.setRunTime((int) (runTimeInNS / 1000000L));
                 Session.getCurrent().getDataManager().logDataRecord(customData);
 
                 if (LOG.isDebugEnabled())
@@ -183,7 +183,7 @@ public final class XltJavaScriptEngine extends JavaScriptEngine
             {
                 final long runTimeInNS = System.nanoTime() - start;
 
-                customData.setRunTime(runTimeInNS / 1000000L);
+                customData.setRunTime((int) (runTimeInNS / 1000000L));
                 Session.getCurrent().getDataManager().logDataRecord(customData);
 
                 if (LOG.isDebugEnabled())
@@ -248,7 +248,7 @@ public final class XltJavaScriptEngine extends JavaScriptEngine
                 final long runTimeInNS = System.nanoTime() - start;
 
                 customData.setName("Compiling " + sourceFileName);
-                customData.setRunTime(runTimeInNS / 1000000L);
+                customData.setRunTime((int) (runTimeInNS / 1000000L));
                 Session.getCurrent().getDataManager().logDataRecord(customData);
 
                 if (LOG.isDebugEnabled())
